pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.7.2"
PyJWT = "^2.8.0"
requests = "^2.31.0"
pika = "^1.3.2"
motor = "^3.4.0"
fastapi = "^0.110.2"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
python-dateutil = "^2.9.0.post0"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.29.0"
pytest = "^8.2.0"
tox = "^4.15.0"
coverage = "^7.5.1"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
